Add a shadow to the menu
This helps make this match the shadow behavior I've added to the map
card, helping give the illusion of depth.

To keep the shadow hidden I had to move the nav container further off
screen. Then when it opened the toggle appeared awkwardly placed on my
iPhone, so on opening I'm shifting it back into position with a
`translateX`.