Shrink menu toggle target size
I accidentally had this spanning the entire vertical space, meaning that
if you tried to click on a pin on the left hand side of the map the menu
would open instead.

Setting `align-self: flex-start;` makes this take up only as much
vertical space as necessary and fit to the top of the cross-axis.